+++ sdk/loadtesting/playwright/src/common/state.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+(globalThis as any).state = (globalThis as any).state || {
+  playwrightServiceEntra: null,
+  customerConfig: null,
+};
+export const state: Record<string, any> = (globalThis as any).state;