+++ sdk/monitor/monitor-opentelemetry-exporter/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Release History
 
-## 1.0.0-beta.32 ()
+## 1.0.0-beta.32 (2025-06-09)
+
+### Features Added
+
+- Added customer-facing statsbeat preview.
 
 ### Other Changes
 
@@ -9,6 +13,8 @@
 - No longer send statsbeat counters when values are zero.
 - Fix statsbeat throttle recording logic.
 - SEMATTRS_ENDUSER_ID is properly added to tags but not to properties in telemetry envelopes.
+- Update network statsbeat to follow a singleton pattern.
+- Stop sending client OS value.
 
 ## 1.0.0-beta.31 (2025-04-16)